Support Brotli decompression (content-encoding: br)
Adds backwards-compatible support for Brotli decompression for versions of Node that natively support Brotli (10.16.0+). For reference, Node LTS is soon to be 14.15.0+ only.

If the version of Node doesn't support Brotli and the request specifies `content-encoding: br`, an HTTP 415 error is thrown, the exact same as before. If Node does support Brotli, it will use Brotli decompression on the provided stream.

A new option called `options.brotli` is supported. This way Brotli and Gzip options are kept separate. Only the `brotli` options are passed to the Brotli decompressor.

Added unit tests for both new and old versions of Node. Added Node 16 to the Travis CI matrix.